add Heartwood.Registry
Some venues/markets/sources will need to be named for ease of access by
bots who subscribe to them. These publishers might also run as processes
themselves. That's why we bring in a simple wrapper around the Registry
module to name processes.

We associate the name of the module with the publisher name as well,
making it easy to call the topics() functions.