add sample Heartwood topography and mechanisms to start it
Topography goes in the configuration files (for now).

Heartwood.Loader reads the config and starts all the children that need
to be started -- that is, all the children with a start_link() function.
These can be venues, markets, sources, and/or bots: Heartwood.Supervisor
starts and manages them all.

We use :rest_for_one at the App level since if the registry crashes,
basically everything else needs to restart, but the Supervisor crashing
doesn't mean the registry should restart. Plus, if the supervisor
crashes we'll need to run the loader again.