take shortcuts with Alpaca partial fills
I noticed Alpaca trade updates have a field for partials and fills
called `qty`, which seems to indicate the quantity filled by that
update. I updated the Ledger logic to use this quantity instead of
having to calculate the quantity filled by that portion every time.

In the future, if there are order updates from venues that don't have
this feature, I could keep track of the last quantity for a particular
order in a GenServer or Agent and do a simple calculation to find the
`qty` equivalent.