fix: cancel orders could delete Ledger's position
When a cancel order comes in, the Ledger assumed that the cancel was on
a new position and deleted the key. However, this is not always the
case. If we opened a position, then canceled the order to close the
position, the Ledger would delete the records from the open position.

This commit ensures that the Ledger only deletes a position on cancel if
it's a new position.