make Polygon's topic/1 available in Data module
This will be needed when we subscribe to new data streams to check on
price updates. To do this, we register Polygon with the Registry under
its provided name, then we look up the name every time we want to call
topic. It's basically the same as how we call specific Alpaca methods
from the generic Venue.