group and normalize inbound Polygon data messages
Polygon data arrives as a list, and there may be different
events/symbols in the list. To address this, group the elements of the
list by topic, then go through each group and broadcast it on the
appropriate channel.

additionally, Polygon data needs to be normalized. For now, just
collect the symbol and price (as a Decimal) of each data point.