add Conditions module and update Production bot to use it
The Conditions module allows orders to store and check various
conditions. You can also store a condition after a delay -- the module
sends a message to the calling process when the condition is added.
Right now, all you can do is check if a price passes a threshold.

Each condition is a two element tuple which looks like: `{condition,
action}`. The condition is what is checked, and the action is what to do
when the condition passes a check.

Conditions that pass a given check are returned as a list of `{id,
action}` tuples.

The Production bot uses the Conditions module to make sure limit orders
get cancelled when the price for that order's symbol exceeds 0.33% in
the wrong direction.