fix Data subscription channel issues
The Xylem channels do not correspond 1:1 with Polygon subscriptions.
Therefore, we need to adjust the Polygon module to subscribe to the
right tickers on the Polygon server, then map those tickers to Xylem
topics.

This also adds in some abstracted Data functions for subscribe/2 and
unsubscribe/2, as well as adds those two callbacks for the Xylem.Data
behaviour. Calling Xylem.Data.{subscribe, unsubscribe} looks for the
right child and delegates subscriptions and unsubscriptions there.

To take advantage of this, Polygon needs to implement subscribe/2 and
unsubscribe/2. I had set up some handle_other/2 callbacks for taking
care of this, so all I needed to do was call send() in the
implementation. I also took care of subscribing to the Xylem topics here
too.